подправил функционал удаления категорий согласно тз

+++ src/service/data-service/category.js
@@ -1,39 +1,20 @@
 'use strict';
 
-const Sequelize = require(`sequelize`);
-
-const Aliase = require(`../models/aliase`);
-
 class CategoryService {
   constructor(sequelize) {
     this._Category = sequelize.models.Category;
     this._ArticleCategory = sequelize.models.ArticleCategory;
+    this._sequelize = sequelize;
   }
 
   async findAll(needCount) {
     if (needCount) {
-      const result = await this._Category.findAll({
-        attributes: [
-          `id`,
-          `name`,
-          [
-            Sequelize.fn(
-                `COUNT`,
-                `*`
-            ),
-            `count`
-          ]
-        ],
-        group: [Sequelize.col(`Category.id`)],
-        order: [`id`],
-        include: [{
-          model: this._ArticleCategory,
-          as: Aliase.ARTICLE_CATEGORIES,
-          attributes: []
-        }]
-      });
+      const result = await this._sequelize.query(`SELECT "Category"."id", "Category"."name",
+      COUNT('*') AS "count" FROM "ArticleCategories" AS "articleCategories"
+      LEFT OUTER JOIN "categories" AS "Category" ON "articleCategories"."CategoryId" = "Category"."id"
+      GROUP BY "Category"."id" ORDER BY "Category"."id";`);
 
-      return result.map((it) => it.get());
+      return result[1].rows;
     } else {
       return await this._Category.findAll({raw: true});
     }
@@ -54,9 +35,17 @@ class CategoryService {
   }
 
   async drop(id) {
-    const deletedRows = await this._Category.destroy({
-      where: {id}
-    });
+    const categoryIdArticles = await this._ArticleCategory.findAll({where: {CategoryId: id}});
+
+    let deletedRows;
+
+    if (categoryIdArticles.length === 0) {
+      deletedRows = await this._Category.destroy({
+        where: {id}
+      });
+    } else {
+      deletedRows = null;
+    }
 
     return !!deletedRows;
   }
